@@ -1,66 +1,42 @@
 {
   "name": "vue-vant-template",
-  "version": "0.1.0",
   "private": true,
+  "version": "0.0.0",
   "scripts": {
-    "start": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint",
-    "stylelint": "stylelint src/**/*.{vue,css,less} --fix"
+    "dev": "vite",
+    "build": "vue-tsc && vite build",
+    "preview": "vite preview"
   },
   "dependencies": {
-    "clipboard": "^2.0.8",
-    "core-js": "^3.6.5",
-    "dayjs": "^1.10.7",
-    "vant": "^2.12.34",
-    "vue": "^2.6.11"
+    "axios": "^1.3.3",
+    "vant": "^4.0.10",
+    "vue": "^3.2.45",
+    "vue-router": "4"
   },
   "devDependencies": {
-    "@commitlint/cli": "^12.1.4",
-    "@commitlint/config-conventional": "^12.1.4",
-    "@vue/cli-plugin-babel": "~4.5.0",
-    "@vue/cli-plugin-eslint": "^4.5.15",
-    "@vue/cli-service": "~4.5.0",
-    "@vue/eslint-config-prettier": "^6.0.0",
-    "amfe-flexible": "^2.2.1",
-    "autoprefixer": "9.8.6",
-    "axios": "^0.24.0",
-    "babel-eslint": "^10.1.0",
-    "babel-plugin-import": "^1.13.3",
-    "compression-webpack-plugin": "6.1.1",
-    "eslint": "^6.7.2",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-vue": "^6.2.2",
-    "husky": "4.3.8",
-    "less": "^4.1.2",
-    "less-loader": "5.0.0",
-    "lint-staged": "^12.0.2",
-    "postcss-pxtorem": "^5.1.1",
-    "prettier": "^2.4.1",
-    "script-ext-html-webpack-plugin": "^2.1.4",
-    "stylelint": "^13.13.1",
-    "stylelint-config-recess-order": "^3.0.0",
-    "stylelint-config-standard": "^22.0.0",
-    "stylelint-less": "^1.0.1",
-    "stylelint-prettier": "^2.0.0",
-    "vue-router": "^3.5.3",
-    "vue-template-compiler": "^2.6.11",
-    "vuex": "^3.6.2",
-    "webpack-bundle-analyzer": "^4.5.0"
+    "@commitlint/cli": "^17.4.4",
+    "@typescript-eslint/eslint-plugin": "^5.52.0",
+    "@typescript-eslint/parser": "^5.52.0",
+    "@vitejs/plugin-vue": "^4.0.0",
+    "commitlint-config-ali": "^0.1.3",
+    "eslint-config-ali": "^14.0.2",
+    "eslint-import-resolver-typescript": "^3.5.3",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-vue": "^9.9.0",
+    "husky": "^8.0.3",
+    "less": "^4.1.3",
+    "less-loader": "^11.1.0",
+    "stylelint": "^15.1.0",
+    "stylelint-config-ali": "^0.4.2",
+    "stylelint-less": "^1.0.6",
+    "typescript": "^4.9.3",
+    "vite": "^4.1.0",
+    "vue-eslint-parser": "^9.1.0",
+    "vue-tsc": "^1.0.24"
   },
   "husky": {
     "hooks": {
-      "pre-commit": "lint-staged",
       "commit-msg": "commitlint -E HUSKY_GIT_PARAMS"
     }
-  },
-  "lint-staged": {
-    "*.{js,vue}": [
-      "vue-cli-service lint",
-      "git add"
-    ],
-    "*.{html,vue,css,less}": [
-      "yarn stylelint"
-    ]
   }
 }
