@@ -1,7 +1,3 @@
 module.exports = {
-  extends: ['stylelint-config-standard', 'stylelint-config-recess-order'],
-  plugins: ['stylelint-less'],
-  rules: {
-    // 校验规则略
-  }
-}
+  extends: 'stylelint-config-ali',
+};
